package_version: Raise PackageNotFound error
Raise `PackageNotFound` error from `package_version` if one or more
specified packages could not be found.

Note that the `available` attribute of the raised `PackageNotFound` object
contains a list of package information dictionaries of the set of
specified packages that **are** available Conda packages.

This is useful, for example, for code to continue processing packages that
**are** found.