feat(run-command): update status indicator by time
Prior to this commit, the waiting status indicator was only updated each
time data was read on either of the `stdout` or `stderr` streams.  In some
cases, this could lead to a significant delay before any update of the
status indicator.

In this commit, update status indicator at time intervals instead.

XXX This commit only implements the described functionality for Python 2.7
(not Python 3.5+).