feat(dev-setup): ensure conda dev versions used
Prior to this commit, the `development_setup` function may have installed
a package that was linked using `conda develop`, causing the package
version to unexpectedly take priority.

In this commit, update `development_setup` to uninstall packages
corresponding to paths added with `conda develop` to ensure the
development versions are used instead.