fix(package-version): add support for `conda>=4.4`
Modify `package_version` function to look up installed package info in
`<prefix>/conda-meta` directory, eliminating dependency on `conda`
executable.

This is useful, for example, with Conda environments created with
`conda>=4.4`, where a link to the root `conda` executable is no longer
created in the `Scripts` directory in the new environment.  In such cases,
it is not possible to locate the root `conda` executable given only the
child environment.

See: ["With conda being a shell function in 4.4+, what's the suggested way to
use conda via scripts when an environment isn't explicitly activated by the
user beforehand?"](https://github.com/conda/conda/issues/7126)